Simplify DataManager
Convert a signal that was just calling functions on private slots of
other objects to a private function that calls them directly, to clarify
the intent. Adjust private slots on those objects to public, to clarify
as well. And convert AvailableModsModel to QStringListModel, as it was
not having any additional functionality.