Modernize all Coordinator connect() calls
This gets checked at compile time, is more efficient, improves finding
matches of existing connections with the tooling, etc. It requires
making public a few functions that were marked as private slots, which
it makes it clearer what's going on (and things are not that well
encapsulated there, with public data members in many cases).