Remove thread usage, and simplify a bit again
A few of the previously removed signals that confused me so badly were
very likely added in order to make the trick of communicating with
objects in different threads via signals and slots. This is fine, but it
wasn't too explicit, and the "worker object" pattern wasn't implemented
properly (with threads leaking and not being stopped, and with extra
indirection for a simple deleteLater). Given that the threads don't seem
necessary (the WeiDU invocations are done via QProcess, with very few
blocking/slow calls), and that after removing the usage, the app seems
to still be responsive, I've opted for commenting out their use.

It's still there in commented out form to be able to quickly test with
and without, and to be able to restore them later if needed.

If they end up being added again, the use of signals will be restricted
from the Controller, who will just use the also thread safe invokeMethod
with a queued connection. Still is a bit frustrating that the call chain
seems a bit obscure, with Coordinator passing the DataManager::getLog
signal to Controller::readLog, who passes it to LogReader::readLog, who
in turn will return a response back to other objects via more signals.